Adding the main file to calc and predict turtle position
Added turtlePrediction.m which is the main loop that reads
the csv data file to track the position of our turtle
It then uses kalmanPredictionOld.m to track and predict
the position of the turtle. It then once per 25 samples
will redraw a graph of the path of the turtle since when
we first started tracking.